Move GitHub API module to utils/github.rb.
